uniform/requirements.txt
nameko
nameko-redis
python-adapters